move common code of PutObjectArgs and UploadObjectArgs
